requirements.txt
diffusers==0.27.2
huggingface_hub==0.23.1
peft==0.11.1
Pillow==10.4.0
protobuf==3.20.3
safetensors==0.4.3
torch==2.3.1
transformers==4.41.1
